# Avoid repeated matching of duplicate impropers