--- a/wrappers/python/simtk/openmm/app/forcefield.py
+++ b/wrappers/python/simtk/openmm/app/forcefield.py
@@ -1998,14 +1998,9 @@ class PeriodicTorsionGenerator(object):
                     if match.k[i] != 0:
                         force.addTorsion(torsion[0], torsion[1], torsion[2], torsion[3], match.periodicity[i], match.phase[i], match.k[i])
         impr_cache = {}
-        # from collections import defaultdict
         for torsion in data.impropers:
             t1, t2, t3, t4 = tatoms = [data.atomType[data.atoms[torsion[i]]] for i in range(4)]
             sig = (t1, t2, t3, t4)
-            # tcount = defaultdict(int)
-            # for t in (t2,t3,t4):
-            #     tcount[t] += 1
-            # sig = (t1, frozenset((atype, count) for atype, count in tcount.items()))
             match = impr_cache.get(sig, None)
             if match == -1:
                 # Previously checked, and doesn't appear in the database