Doing some profiling of simulation startup in a large forcefield with 2922 different improper definitions, outside of `new_Context()` the repeated calls to `_matchImproper()` in `PeriodicTorsionGenerator` are far and away the most expensive (3.5 seconds for a 229-residue protein vs. about 200ms for the next-most-expensive call). By caching the improper definitions (or lack thereof) for signatures that have already been seen, the number of calls to `_matchImproper()` is reduced from ~2500 to 92, bringing the time cost down to 125 ms.

Incidentally, the `createForce()` methods in `RBTorsionGenerator` and `CustomTorsionGenerator` are near-identical copies of that in `PeriodicTorsionGenerator`. Looks like these could be quite easily refactored as subclasses of a generic base class?