@tristanic 2019-02-26
May as well do the same for the proper torsion search - the gain is much smaller than for impropers, but not negligible and I can't really see any harm in it. Some real-world timings using `prun` on my standard test case:

No optimisations:
```
823829 function calls (799340	 primitive calls) in 9.835 seconds
								
Ordered by: internal time					
								
ncalls	tottime	percall	cumtime	percall	filename:lineno(function)			
1	4.542	4.542	4.542	4.542	{built-in	method	simtk.openmm._openmm.new_Context}	
3425	3.944	0.001	3.944	0.001	forcefield.py:1650(_matchImproper)			
1	0.239	0.239	4.233	4.233	forcefield.py:1962(createForce)			
1	0.148	0.148	0.148	0.148	{built-in	method	simtk.openmm._openmm.new_Continuous3DFunction}	
```

Cached improper search:
```
823713 function calls (799224 primitive calls) in 5.379 seconds
							
Ordered by: internal time				
							
ncalls	tottime	percall	cumtime	percall	filename:lineno(function)		
1	3.963	3.963	3.963	3.963	{built-in	method	simtk.openmm._openmm.new_Context}
1	0.223	0.223	0.386	0.386	forcefield.py:1962(createForce)		
1	0.152	0.152	0.152	0.152	{built-in	method	simtk.openmm._openmm.new_Continuous3DFunction}
92	0.126	0.001	0.126	0.001	forcefield.py:1650(_matchImproper)		
```

Cached impropers and propers:
```
832593 function calls (808104 primitive calls) in 5.291 seconds
								
Ordered by: internal time					
								
ncalls	tottime	percall	cumtime	percall	filename:lineno(function)			
1	3.904	3.904	3.904	3.904	{built-in	method	simtk.openmm._openmm.new_Context}	
1	0.211	0.211	0.211	0.211	{built-in	method	simtk.openmm._openmm.new_Continuous3DFunction}	
92	0.127	0.001	0.127	0.001	forcefield.py:1650(_matchImproper)			
1	0.062	0.062	0.229	0.229	forcefield.py:1962(createForce)			



```