@tristanic 2019-02-26
Ouch - two different errors in my original attempt. One: my attempt at a signature for an improper wasn't sufficient (couldn't distinguish between `(type1, type2, type2, type3)` and `(type1, type2, type3, type3)`), and two: I'd badly misunderstood the contents of the `match` tuple. 

Something I'm still a little surprised about: if I uncomment the alternative signature format:
```
            tcount = defaultdict(int)
            or t in (t2,t3,t4):
                tcount[t] += 1
            sig = (t1, frozenset((atype, count) for atype, count in tcount.items()))
```
... then the test fails. With the current ordered version (`sig = (t1, t2, t3, t4`) I get 333 calls to `_matchImproper()` vs. 94 with the alternative signature.

My question is, is this a bug? `_matchImproper()` works through all permutations of the ordering of positions 2, 3 and 4. If there are wildcard matches but no exact matches, it will return the first wildcard match it hits. If there *are* exact matches, it will return the last exact match, and otherwise it will return None. So in at least some cases the improper definition that gets chosen is going to depend on the order in which the code walks through the atom order permutations. That seems out of step with the intention of the code.