@tristanic 2019-02-26
OK, I can see that ordering of the improper atoms *does* matter, albeit in ways I still don't fully understand. With the `frozenset`-based signature, somewhere approaching half the impropers come out with the wrong handedness. Something to do with the "bizarre" ordering rules mentioned in the code? In any case, the version in the last commit should be safe (the same four atom types in order should *always* return the same result), and still gets a ~10-fold speedup for my test case (and I would expect substantially bigger speedups for larger models).