@tristanic 2019-02-26
Some timings on a much larger model (6g2j *sans* lipids - 131k non-solvent atoms)

Original:
```
         31682943 function calls (30747723 primitive calls) in 181.350 seconds

   Ordered by: internal time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
   133061  147.153    0.001  147.162    0.001 forcefield.py:1650(_matchImproper)
        1    2.623    2.623  160.313  160.313 forcefield.py:1023(createSystem)
        1    2.070    2.070  150.859  150.859 forcefield.py:1962(createForce)
        1    1.804    1.804    1.804    1.804 {built-in method simtk.openmm._openmm.new_Context}
```

New:
```
         32094332 function calls (31159112 primitive calls) in 35.893 seconds

   Ordered by: internal time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    2.590    2.590   13.679   13.679 forcefield.py:1023(createSystem)
        1    1.948    1.948    4.232    4.232 forcefield.py:1962(createForce)
1365366/443879    1.760    0.000    4.818    0.000 copy.py:132(deepcopy)
        1    1.611    1.611    1.611    1.611 {built-in method simtk.openmm._openmm.new_Context}
	... 
     562    0.632    0.001    0.632    0.001 forcefield.py:1650(_matchImproper)
```