@peastman 2019-02-26
Impressive!  I'll go through the code shortly.