@peastman 2019-02-26
I'm a bit confused by the line

```python
sig = frozenset((sig, sig[:-1]))
```

Could you explain the logic behind that?