@peastman 2019-02-27
If `sig` is A-B-C-D, then `sig[:-1]` is A-B-C.  Did you mean to write `sig[::-1]`?