@tristanic 2019-02-27
Ah, damn - yes, I did.


 

 

> On 26 Feb 2019, at 23:19, peastman <notifications@github.com> wrote:
> 
> I'm a bit confused by the line
> 
> sig = frozenset((sig, sig[:-1]))
> Could you explain the logic behind that?
> 
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub, or mute the thread.